# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Jewelry Store REST API",
    "description": "REST API theo database.sql 25 bảng. Đặt hàng COD; phân trang bắt đầu từ 0. KHACH_HANG, NHAN_VIEN, QUAN_LY. Xem README/API.md để biết phân quyền và luồng nghiệp vụ.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/admin/contents/{id}": {
      "put": {
        "tags": [
          "content-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: QUAN_LY. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "updateContent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseContentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "QUAN_LY"
        ]
      },
      "delete": {
        "tags": [
          "content-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: QUAN_LY. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "deleteContent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "QUAN_LY"
        ]
      }
    },
    "/api/admin/banners/{id}": {
      "put": {
        "tags": [
          "content-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: QUAN_LY. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "updateBanner",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BannerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBannerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "QUAN_LY"
        ]
      },
      "delete": {
        "tags": [
          "content-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: QUAN_LY. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "deleteBanner",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "QUAN_LY"
        ]
      }
    },
...
    "/api/admin/contents": {
      "get": {
        "tags": [
          "content-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: QUAN_LY. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "adminContents",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "FAQ",
                "CHINH_SACH",
                "GIOI_THIEU",
                "LIEN_HE",
                "TRANG_CHU"
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseContentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "QUAN_LY"
        ]
      },
      "post": {
        "tags": [
          "content-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: QUAN_LY. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "createContent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseContentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "QUAN_LY"
        ]
      }
    },
    "/api/admin/banners": {
      "get": {
        "tags": [
          "content-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: QUAN_LY. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "adminBanners",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseBannerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "QUAN_LY"
        ]
      },
      "post": {
        "tags": [
          "content-controller"
        ],
        "description": "Yêu cầu JWT. Vai trò: QUAN_LY. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "createBanner",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BannerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseBannerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "QUAN_LY"
        ]
      }
    },
...
    "/api/contents": {
      "get": {
        "tags": [
          "content-controller"
        ],
        "description": "Public. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "contents",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "FAQ",
                "CHINH_SACH",
                "GIOI_THIEU",
                "LIEN_HE",
                "TRANG_CHU"
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseContentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [],
        "x-roles": []
      }
    },
    "/api/contents/{slug}": {
      "get": {
        "tags": [
          "content-controller"
        ],
        "description": "Public. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "content",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseContentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [],
        "x-roles": []
      }
    },
...
    "/api/banners": {
      "get": {
        "tags": [
          "content-controller"
        ],
        "description": "Public. Response dùng ApiResponse; page bắt đầu từ 0, size tối đa 100. ",
        "operationId": "banners",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseBannerResponse"
                }
              }
            }
          },
          "400": {
            "description": "Request hoặc validation không hợp lệ"
          },
          "401": {
            "description": "Chưa đăng nhập hoặc JWT hết hiệu lực"
          },
          "403": {
            "description": "Không đủ quyền hoặc không sở hữu dữ liệu"
          },
          "404": {
            "description": "Không tìm thấy"
          },
          "409": {
            "description": "Xung đột dữ liệu, tồn kho hoặc trạng thái"
          },
          "500": {
            "description": "Lỗi máy chủ"
          }
        },
        "security": [],
        "x-roles": []
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {}
        }
      },
...
      "ContentRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "FAQ",
              "CHINH_SACH",
              "GIOI_THIEU",
              "LIEN_HE",
              "TRANG_CHU"
            ]
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "slug": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0,
            "pattern": "[a-z0-9]+(?:-[a-z0-9]+)*"
          },
          "content": {
            "type": "string",
            "maxLength": 100000,
            "minLength": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "HIEN_THI",
              "AN"
            ]
          }
        },
        "required": [
          "content",
          "slug",
          "status",
          "title",
          "type"
        ]
      },
      "ApiResponseContentResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ContentResponse"
          }
        }
      },
      "ContentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string",
            "enum": [
              "FAQ",
              "CHINH_SACH",
              "GIOI_THIEU",
              "LIEN_HE",
              "TRANG_CHU"
            ]
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "HIEN_THI",
              "AN"
            ]
          }
        }
      },
      "BannerRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          },
          "image": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 0
          },
          "link": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 0
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "status": {
            "type": "string",
            "enum": [
              "HIEN_THI",
              "AN"
            ]
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "image",
          "status"
        ]
      },
      "ApiResponseBannerResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/BannerResponse"
          }
        }
      },
      "BannerResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "link": {
            "type": "string"
          },
          "sortOrder": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string",
            "enum": [
              "HIEN_THI",
              "AN"
            ]
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ApiResponsePageResponseContentResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResponseContentResponse"
          }
        }
      },
      "PageResponseContentResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContentResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ApiResponsePageResponseBannerResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResponseBannerResponse"
          }
        }
      },
      "PageResponseBannerResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BannerResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}